+++ front/src/App.jsx
@@ -6,7 +6,7 @@ import {Route, Routes} from "react-router-dom";
 //Componentes
 import {Home} from "./components/home/Home"
 import { Catalogo } from './components/catalogo/catalogo';
-import BookList from './components/home/BookList'
+
 
 function App() {
   return (
@@ -14,7 +14,7 @@ function App() {
       <Routes>
         <Route exact path="/home" element={<Home/>}/>
         <Route exact path="/catalogo" element={<Catalogo/>}/>
-        <Route exact path="/books" element={<BookList />}/>
+      
       </Routes>
     </>
   )
